faas/function_mode/func3/requirements.txt
requests
torch-cpu
torchvision-cpu
